% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.vtcustom}
\alias{coef.vtcustom}
\title{Coefficient Method for vtcustom Objects.}
\usage{
\method{coef}{vtcustom}(object, ...)
}
\arguments{
\item{object}{An object of class \code{vtcustom}.}

\item{...}{Currently not implemented}
}
\description{
Returns coefficients from an object of class \code{vtcustom}.  This allows users to build a
list with estimates and a variance covariance matrix.  If the object's class is set to \code{vtcustom}
and the object has an element \code{coef}, this coefficient method will return the vector of estimates.
For maximum usability, the entries of the coefficient vector should be named.
}
