% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_letters}
\alias{get_letters}
\title{Get Letters for Multiple Comparisons}
\usage{
get_letters(x, linfct, ...)
}
\arguments{
\item{x}{An object that is compatible with the \code{glht()} function}

\item{linfct}{The linear function contrasts to be used in the multiple comparisons.}

\item{...}{Arguments to be passed down to \code{glht()} or \code{summary.glht()}.  Arguments for the former include \code{coef.}, \code{vcov.}, and \code{df}.
Arguments for the latter include \code{test} and a host of others.  See the help file for \code{summary.glht} for examples.}
}
\description{
Gets the letter matrix for a compact letter display.  This can be passed to the \code{letter_plot()} function from the \code{psre} package to produce plots of
confidence intervals with a letter display.
}
