% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_segs}
\alias{make_segs}
\title{Make Reference Segments}
\usage{
make_segs(.data, vdt = 0.02, ...)
}
\arguments{
\item{.data}{Data to be used in the calculating segments.  The data should contain the variables \code{lwr} and \code{upr} for the lower and upper confidence bounds, respectively.}

\item{vdt}{Visual difficulty threshold, see the details of \code{help(plot.viztest)} for an explanation of the parameter.}

\item{...}{Other arguments to be passed down.  Currently not implemented.}
}
\value{
A data frame with variables \code{stim_start} and \code{stim_end} identifying the starting and ending stimuli for the segments as well as \code{bound_start} and \code{bound_end} which give the relevant value of the reference bound.  There is also a variable called \code{ambiguous} which indicates if any comparisons with that bound are ambiguous.
}
\description{
Makes reference segments for the \code{plot.viztest}.  These segments run along the upper bound of a stimulus from the stimulus in question to the last comparison stimulus whose lower bound is smaller than the upper bound of the stimulus in question.
}
