% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.gg_forest}
\alias{plot.gg_forest}
\title{Plotting method for \code{gg_forest} objects.}
\usage{
\method{plot}{gg_forest}(x, ..., widths = c(1.4, 1))
}
\arguments{
\item{x}{An object of class \code{"gg_foest"}, produces with the \code{gg_forest()} function.}

\item{...}{Other arguments passed to \code{theme()} to control the appearance of the combined plot.
For example, you might want to use \code{legend.position = "none"} if you don't want a legend for the CI colors.}

\item{widths}{Numeric vector of length 2 giving the relative widths of the table and forest plot components.
The default is \code{c(1.4, 1)}, which gives the table slightly more width than the forest plot.}
}
\value{
A ggplot object combining the forest plot and table components.
}
\description{
Uses patchwork to combine the forest plot and table components.
}
