% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vcov.vtcustom}
\alias{vcov.vtcustom}
\title{Variance-Covariance Method for vtcustom Objects.}
\usage{
\method{vcov}{vtcustom}(object, ...)
}
\arguments{
\item{object}{An object of class \code{vtcustom}.}

\item{...}{Currently not implemented}
}
\description{
Returns the variance-covariance matrix from an object of class \code{vtcustom}.  This allows users to build a
list with estimates and a variance covariance matrix.  If the object's class is set to \code{vtcustom}
and the object has an element \code{vcov}, this coefficient method will return the variance-covariance matrix.
The matrix may have row and column names, but they will have no effect on the result.
}
