% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{viz_compare}
\alias{viz_compare}
\title{Compare two output folders}
\usage{
viz_compare(output_a, output_b)
}
\arguments{
\item{output_a, output_b}{folders to compare that contain output images}
}
\description{
Compare two output folders
}
