% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{viztest.default}
\alias{viztest.default}
\title{Default method for viztest function}
\usage{
\method{viztest}{default}(
  obj,
  test_level = 0.05,
  range_levels = c(0.25, 0.99),
  level_increment = 0.01,
  adjust = c("none", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"),
  include_intercept = FALSE,
  include_zero = TRUE,
  ...
)
}
\arguments{
\item{obj}{A model object (or any object) where \code{coef()} and \code{vcov()} return estimates of coefficients and sampling variability.}

\item{test_level}{The type I error rate of the pairwise tests.}

\item{range_levels}{The range of confidence levels to try.}

\item{level_increment}{Step size of increase between the values of \code{range_levels}.}

\item{adjust}{Multiplicity adjustment to use when calculating the p-values for the pairwise tests.}

\item{include_intercept}{Logical indicating whether the intercept should be included in the tests, defaults to \code{FALSE}.}

\item{include_zero}{Should univariate tests at zero be included, defaults to \code{TRUE}.}

\item{...}{Other arguments, currently not implemented.}
}
\description{
Default method for viztest function
}
