% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viztest.R
\name{viztest}
\alias{viztest}
\title{Visual Test Package Examples}
\usage{
viztest(pkg = ".", output_dir = file.path("viztest",
  paste0(devtools::as.package(pkg)$package, "-",
  devtools::as.package(pkg)$version)), ..., delay = 2, fig.width = 8,
  fig.height = 6, vwidth = 992, vheight = 744, test = TRUE,
  run = FALSE, stomp = FALSE, save_individual = FALSE)
}
\arguments{
\item{pkg}{Package to load with devtools}

\item{output_dir}{Save directory. Defaults to "viztest/PKG-VERSION"}

\item{...}{ignored}

\item{delay}{amount of delay to use before capturing}

\item{fig.width, fig.height}{Figure width and height in inches}

\item{vwidth, vheight}{Screenshot size in pixels}

\item{test}{If \code{TRUE}, code in \code{\\donttest{}} will be commented out.
If \code{FALSE}, code in \code{\\testonly{}} will be commented out.}

\item{run}{If \code{TRUE}, code in \code{\\dontrun{}} will be commented out.}

\item{stomp}{If \code{TRUE}, the prior output folder will be completely deleted before executing}

\item{save_individual}{If \code{TRUE}, the individual knitr files used to produce the images will be saved in the output folder}
}
\description{
Execute all package examples in independent knitr files to capture images of each visual item produced, such as plots, htmlwidgets, shiny application objects.
}
\examples{
\dontrun{
# R session running in local R package folder
viztest(stomp = TRUE)
}
}
