% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viztest.R
\name{viztest}
\alias{viztest}
\title{Visually Test Package Examples}
\usage{
viztest(pkg = ".", old_pkg = paste0(devtools::as.package(pkg)$package),
  output_dir = file.path(paste("viztest", devtools::as.package(pkg)$package,
  devtools::as.package(pkg)$version, sep = "-")), ..., delay = 2,
  fig.width = 8, fig.height = 6, vwidth = 992, vheight = 744,
  test = TRUE, run = FALSE, stomp = FALSE, cache = TRUE,
  save_individual = TRUE, resize = TRUE, browse = TRUE)
}
\arguments{
\item{pkg}{Local R package to load with devtools}

\item{old_pkg}{R package name description.  This should either be a CRAN or GitHub name to
work with \code{devtools::\link[devtools]{install_cran}}
or \code{devtools::\link[devtools]{install_github}} respectively.}

\item{output_dir}{Output directory. Defaults to "viztest-PKG-VERSION"}

\item{...}{parameters sent to \code{\link{viz_compare}}}

\item{delay}{Amount of delay to use before capturing}

\item{fig.width, fig.height}{Figure width and height in inches}

\item{vwidth, vheight}{Screenshot size in pixels}

\item{test}{If \code{TRUE}, code in \code{\\donttest{}} will be commented out.
If \code{FALSE}, code in \code{\\testonly{}} will be commented out.}

\item{run}{If \code{TRUE}, code in \code{\\dontrun{}} will be commented out.}

\item{stomp}{If \code{TRUE}, allows \code{viztest} to reexecute in an existing output directory}

\item{cache}{If \code{TRUE}, the local R package examples will be cached with knitr for faster
execution.  The \code{old_pkg} will always be cached as it's source code does not change.}

\item{save_individual}{If \code{TRUE}, individual example knitr files will be saved}

\item{resize}{\code{resize} parameter supplied to \code{webshot::\link[webshot]{resize}}.
Defaults (\code{TRUE}) to shrink each dimension to 50%. Shrinking will not alter
the original files.}

\item{browse}{If \code{TRUE}, diff.html in \code{output_dir} will be opened}
}
\description{
Execute all package examples in independent knitr files to capture images
of each visual item produced, such as plots, htmlwidgets, shiny application objects.
}
\examples{
\dontrun{
# R session running in local R package folder
viztest(stomp = TRUE)
}
}
